@@ -25,7 +25,7 @@ SOFTWARE.
 from __future__ import annotations
 
 from collections.abc import Generator
-from typing import Any, Optional
+from typing import Any
 
 from .abc import Hashable, ReconstructAble
 from .asset import Asset
@@ -156,7 +156,7 @@ class POI(Hashable, ReconstructAble[dict[str, Any], HTTPClientT]):
         super().__init__(data=data, http=http)
 
         self.id: str = data["id"]
-        self.name: Optional[str] = data.get("name")
+        self.name: str | None = data.get("name")
         self.location: POILocation[HTTPClientT] = POILocation(data=data["location"], http=http)
 
 
