@@ -24,7 +24,7 @@ SOFTWARE.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any
 
 from .abc import Hashable, ReconstructAble
 from .asset import Asset
@@ -59,10 +59,10 @@ class PlaylistImages(ReconstructAble[dict[str, Any], HTTPClientT]):
         super().__init__(data=data, http=http)
 
         _showcase = data.get('showcase')
-        self.showcase: Optional[Asset[HTTPClientT]] = _showcase and Asset(url=_showcase, http=http)
+        self.showcase: Asset[HTTPClientT] | None = _showcase and Asset(url=_showcase, http=http)
 
         _mission_icon = data.get('missionIcon')
-        self.mission_icon: Optional[Asset[HTTPClientT]] = _mission_icon and Asset(url=_mission_icon, http=http)
+        self.mission_icon: Asset[HTTPClientT] | None = _mission_icon and Asset(url=_mission_icon, http=http)
 
 
 class Playlist(Hashable, ReconstructAble[dict[str, Any], HTTPClientT]):
@@ -144,11 +144,11 @@ class Playlist(Hashable, ReconstructAble[dict[str, Any], HTTPClientT]):
         super().__init__(data=data, http=http)
         self.id: str = data['id']
         self.name: str = data['name']
-        self.sub_name: Optional[str] = data.get('subName')
-        self.description: Optional[str] = data.get('description')
+        self.sub_name: str | None = data.get('subName')
+        self.description: str | None = data.get('description')
 
-        self.game_type: Optional[str] = data.get('gameType')  # TODO: Make this into an enum
-        self.rating_type: Optional[str] = data.get('ratingType')
+        self.game_type: str | None = data.get('gameType')  # TODO: Make this into an enum
+        self.rating_type: str | None = data.get('ratingType')
 
         self.min_players: int = data['minPlayers']
         self.max_players: int = data['maxPlayers']
@@ -164,7 +164,7 @@ class Playlist(Hashable, ReconstructAble[dict[str, Any], HTTPClientT]):
         self.accumulate_to_profile_stats: bool = data['accumulateToProfileStats']
 
         _images = data.get('images')
-        self.images: Optional[PlaylistImages[HTTPClientT]] = _images and PlaylistImages(data=_images, http=http)
+        self.images: PlaylistImages[HTTPClientT] | None = _images and PlaylistImages(data=_images, http=http)
 
         self.gameplay_tags: list[str] = get_with_fallback(data, 'gameplayTags', list)
         self.path: str = data['path']
